requirement.txt
pillow
pycrypto
numpy
aiohttp
aiodns
tqdm
